+++ Final Prep/RSA.c
@@ -4,17 +4,24 @@
 int modulus(int m, int e, int n) {
     // m is base, e is exponent, n is modulus
 
-    int rm = m; // Return value
+    int rm; // Return value
 
-    for (int i = 1 << ((int) log2(e) - 1); i > 0; i >>= 1) {
-        // Set i to 2^(length-1) and divide it by 2 every loop
-        // (int) log2(e) returns the length (base 0) of the binary representation of e
-        // Length (base 0, i.e. start counting from 0) = Length + 1 (in base 1)
-        rm *= rm;
-        rm %= n;
-        if (e & i) { // Checking if the current bit in the exponent is a 1 or not
-            rm *= m;
+    if (e == 0) {
+        rm = 1;
+    } else if (e == 1) {
+        rm = m % n;
+    } else {
+        rm = m;
+        for (int i = 1 << ((int) log2(e) - 1); i > 0; i >>= 1) {
+            // Set i to 2^(length-1) and divide it by 2 every loop
+            // (int) log2(e) returns the length (base 0) of the binary representation of e
+            // Length (base 0, i.e. start counting from 0) = Length + 1 (in base 1)
+            rm *= rm;
             rm %= n;
+            if (e & i) { // Checking if the current bit in the exponent is a 1 or not
+                rm *= m;
+                rm %= n;
+            }
         }
     }
     return rm;
@@ -53,7 +60,7 @@ int main() {
                     printf("e = ");
                     scanf("%d", &e);
                     d = modMulInv(e, (p - 1) * (q - 1));
-                    printf("d = %d\n",d);
+                    printf("d = %d\n", d);
                     verify = 0;
                     break;
                 case 'd':
@@ -61,7 +68,7 @@ int main() {
                     printf("d = ");
                     scanf("%d", &d);
                     e = modMulInv(d, (p - 1) * (q - 1));
-                    printf("e = %d\n",e);
+                    printf("e = %d\n", e);
                     verify = 0;
                     break;
                 default:
@@ -89,10 +96,10 @@ int main() {
                 scanf("%d", &m);
 
                 printf("Encrypting...\n");
-                c = modulus(m,e,p*q);
-                printf("Ciphertext (C): %d\n",c);
+                c = modulus(m, e, p * q);
+                printf("Ciphertext (C): %d\n", c);
                 printf("Decrypting...\n");
-                printf("Plaintext (M): %d\n", modulus(c,d,p*q));
+                printf("Plaintext (M): %d\n", modulus(c, d, p * q));
 
                 verify = 0;
                 break;
@@ -105,10 +112,10 @@ int main() {
                 scanf("%d", &c);
 
                 printf("Decrypting...\n");
-                m = modulus(c,d,p*q);
+                m = modulus(c, d, p * q);
                 printf("Plaintext (M): %d\n", m);
                 printf("Encrypting...\n");
-                printf("Ciphertext (C): %d\n",modulus(m,e,p*q));
+                printf("Ciphertext (C): %d\n", modulus(m, e, p * q));
 
                 verify = 0;
                 break;
